add SiteTypeService refactor
